requirements.txt
attrs==21.4.0
autobahn==21.11.1
Automat==20.2.0
binance-connector==1.10.0
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.10
coinapi.rest.v1==1.3
constantly==15.1.0
cryptography==36.0.1
hyperlink==21.0.0
idna==3.3
incremental==21.3.0
parse==1.19.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
pyOpenSSL==21.0.0
requests==2.27.1
service-identity==21.1.0
six==1.16.0
Twisted==21.7.0
txaio==21.2.1
typing_extensions==4.0.1
urllib3==1.26.8
zope.interface==5.4.0
